fix: extract filterSubgraph to utility module and guard JSON.parse
Move filterSubgraph from execution-subgraph.tsx to a dedicated utility
module to satisfy react-refresh/only-export-components lint rule. Guard
attributeSchema JSON.parse in account-types page to prevent crash on
malformed data.
